Move implementations from headers to source files
